Move Flatten to its own module
